Messagebox and control points added

+++ app.py
@@ -1,18 +1,19 @@
 from tkinter import *
-from tkinter import filedialog
+from tkinter import filedialog, messagebox
 from modules.gui import MrCleaner
+from modules.report_gen import report_generator
 from modules.multiples import base_converter
 import os, shutil
 
 
 root = Tk()
-root.title("MrCleaner 1.0")
+root.title("MrCleaner ver:1.0")
 root.geometry("700x500+300+100")
 root.minsize(400, 300)
 
 #   IMPORTING VARIABLES FROM GUI    #
 gui = MrCleaner(root)
-input = gui.input
+input_ = gui.input
 list_box = gui.list_box
 ctrl_del = gui.ctrl_del
 ctrl_rep = gui.ctrl_rep
@@ -25,20 +26,24 @@ sizes = []
 
 #   FUNCTIONS   #
 def get_directory():
-
-    directory = filedialog.askdirectory(title="Select messy directory",
+    directory = filedialog.askdirectory(title="Select directory",
                                         mustexist=True)
-    
-    input.delete(0, END)
-    input.insert(END, directory)
+    input_.delete(0, END)
+    input_.insert(END, directory)
 
-    locate_duplicated_files()
+    if directory:
+        msg = '''NOTE: Depending on how deep your subdirectories is,\
+it my take a while and or seem to be frozen, so please, be patient and do not interrupt!'''
+        messagebox.showinfo('Message', msg)
+        locate_duplicated_files()
 
 
 def locate_duplicated_files():
     global duplicated_files, sizes
+    duplicated_files = []
+    sizes = []
 
-    directory = input.get()
+    directory = input_.get()
     
     found_files = []
     files_path = []
@@ -65,17 +70,18 @@ def locate_duplicated_files():
 
     list_box.delete(0, END)
 
-    header = "====================== DUPLICATED FILES ======================"
+    if len(duplicated_files) != 0:
+        list_box.insert(END, "{:=^79}".format(' DUPLICATED FILES '))
+        list_box.insert(END, '')
 
-    list_box.insert(END, header)
-
-    for f in range(len(duplicated_files)):
-        file = f"{duplicated_files[f]} Size: {base_converter(files_sizes[f])}"
-        list_box.insert(END, file)
+        for f in range(len(duplicated_files)):
+            file = f"{duplicated_files[f]} Size: {base_converter(files_sizes[f])}"
+            list_box.insert(END, file)
+    else:
+        messagebox.showinfo('Message', 'No duplicated file found!')
 
 
 def extract_file_from_duplicated(dub_files, sizes):
-
     only_names = []
     file_path = []
     file_names = []
@@ -89,48 +95,57 @@ def extract_file_from_duplicated(dub_files, sizes):
             file_names.append(only_names[f])
             file_path.append(dub_files[f])
             file_size.append(sizes[f])
-
     return file_path, file_size
-
-
 
 
 def remove_duplicated_files():
     global duplicated_files, sizes
-
-    destiny_dir = filedialog.askdirectory(title="Choose backup directory", 
-                                        mustexist=True)
-    file_path, file_size = extract_file_from_duplicated(duplicated_files, sizes)
-
-    count = 0
-    try:
-        for i in range(len(duplicated_files)): 
-            if duplicated_files[i] in file_path:
-                shutil.copy(src=duplicated_files[i], dst=destiny_dir)
+    
+    if len(duplicated_files) != 0:
+        file_path, file_size = extract_file_from_duplicated(duplicated_files, sizes)
+        count = 0
         
-            if ctrl_del.get() == 1:
-                os.remove(duplicated_files[i])
-                count += 1
-    except:
-        pass
+        if ctrl_del.get() == 1:
+            msg = 'You are about to delete files! Do you wnat to continue?'
+            answer = warning_on_delete = messagebox.askquestion('Confirmation', msg)
 
-    if ctrl_rep.get() == 1:
-        report_generator(duplicated_files, sizes)
+            if answer == 'yes':
+                destiny_dir = filedialog.askdirectory(title="Choose backup directory", 
+                                                        mustexist=True)
+                try:
+                    for i in range(len(duplicated_files)): 
+                        if duplicated_files[i] in file_path:
+                            shutil.copy(src=duplicated_files[i], dst=destiny_dir)
+                        os.remove(duplicated_files[i])
+                        count += 1
+                except:
+                    pass            
+                if answer == 'yes':
+                    list_box.delete(0, END)
+                    list_box.insert(0, '{:=^79}'.format(' REMAINING FILES '))
+                    list_box.insert(END, '')
 
-    list_box.delete(0, END)
+                    for j in range(len(file_path)):
+                        item = f"{file_path[j]}  {base_converter(file_size[j])}"
+                        list_box.insert(END, item)
+            else:
+                messagebox.showerror('Message', 'Aborting!')
+        else:    
+            pass            
+        if ctrl_rep.get() == 1:
+            report_generator(duplicated_files, sizes, base_converter)
+            
+        if ctrl_del.get() == 0 and ctrl_rep.get() == 0:
+            msg = 'Please, you must choose an option!'
+        else:
+            msg = 'Success!'
+            list_box.delete(0, END)
+        messagebox.showinfo('Message', msg)
+    else:
+        path = input_.get()
+        if not path:    
+            messagebox.showinfo('Message', 'Please, select directory!')
 
-    for j in range(len(file_path)):
-        list_box.insert(END, (file_path[j], base_converter(file_size[j])))
-
-
-
-
-def report_generator(dub_files, sizes):
-
-    with open('report.txt', 'w') as rep:
-        rep.write('{:~^60}\n\n'.format('Duplicated files found'))
-        for l in range(len(dub_files)):
-            rep.write(f'{l+1} - {dub_files[l]}  -  {base_converter(sizes[l])}\n')
